third_party/WebKit/LayoutTests/fast/events/drag-and-drop-autoscroll-mainframe-expected.txt

state=START
state=NE
PASS document.body.scrollLeft > 0 is true
PASS !document.body.scrollTop is true
state=SE
PASS document.body.scrollLeft > 0 is true
PASS document.body.scrollTop > 0 is true
state=SW
PASS document.body.scrollLeft < lastScrollLeft is true
PASS document.body.scrollTop > 0 is true
state=NW
PASS document.body.scrollLeft <= lastScrollLeft is true
PASS document.body.scrollTop < lastScrollTop is true
PASS successfullyParsed is true

TEST COMPLETE
